pyproject.toml
[tool.poetry]
name = "masterarbeit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.2.4"
scikit-learn = "^0.24.2"
seaborn = "^0.11.1"
jupyter = "^1.0.0"
keras = "^2.4.3"
tensorflow = "^2.5.0"
plotly = "^4.14.3"
python-box = "^5.3.0"
tensorflow-federated = "^0.19.0"
pre-commit = "^2.13.0"
black = "^21.6b0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
